Feature/forest cell renderer and spacial index

+++ libs/sharedmodels/src/main/java/com/recom/dto/map/topography/HeightMapDescriptorDto.java
@@ -25,7 +25,7 @@ public class HeightMapDescriptorDto implements Serializable {
 
     @Schema
     @JsonProperty()
-    private Float stepSize;
+    private Integer stepSize;
 
     @Schema
     @JsonProperty()
